+++ starlight-core/src/main/java/org/atcraftmc/starlight/framework/module/BukkitAutoRegManager.java
@@ -1,8 +1,10 @@
 package org.atcraftmc.starlight.framework.module;
 
 import me.gb2022.gluon.Registrations;
+import org.atcraftmc.starlight.core.VisualScoreboardService;
 import org.atcraftmc.starlight.core.platform.BukkitUtil;
 import org.atcraftmc.starlight.core.view.PlayerUIService;
+import org.atcraftmc.starlight.core.view.ScoreboardTrackingStateCallback;
 import org.atcraftmc.starlight.core.view.UITrackingStateCallback;
 import org.atcraftmc.starlight.util.PluginAutoRegManager;
 import org.bukkit.event.Listener;
@@ -16,6 +18,14 @@ public final class BukkitAutoRegManager extends PluginAutoRegManager {
             i.detach(Registrations.SERVER_EVENT, (o) -> BukkitUtil.unregisterEventListener(((Listener) o)));
             i.attach(PlayerUIService.TRACKING, (o) -> PlayerUIService.TRACKER.attachCallback(((UITrackingStateCallback) o)));
             i.detach(PlayerUIService.TRACKING, (o) -> PlayerUIService.TRACKER.detachCallback(((UITrackingStateCallback) o)));
+            i.attach(
+                    VisualScoreboardService.TRACKING,
+                    (o) -> VisualScoreboardService.instance().attachCallback(((ScoreboardTrackingStateCallback) o))
+            );
+            i.detach(
+                    VisualScoreboardService.TRACKING,
+                    (o) -> VisualScoreboardService.instance().detachCallback(((ScoreboardTrackingStateCallback) o))
+            );
         });
     }
 }