+++ starlight-core/src/main/java/org/atcraftmc/starlight/util/CachedInfo.java
@@ -7,6 +7,7 @@ import org.bukkit.event.EventHandler;
 import org.bukkit.event.Listener;
 import org.bukkit.event.player.PlayerJoinEvent;
 import org.bukkit.event.player.PlayerLoginEvent;
+import org.bukkit.event.player.PlayerQuitEvent;
 
 import java.util.List;
 
@@ -45,7 +46,7 @@ public interface CachedInfo {
         }
 
         @EventHandler
-        public void onPlayerLeave(PlayerLoginEvent event) {
+        public void onPlayerLeave(PlayerQuitEvent event) {
             refreshPlayerNames();
         }
     }